rename generatePreFilledLink to getPreFilledLink
