+++ src/components/UpdateResourceButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Button } from '@material-ui/core';
 
 import { TResource } from '../types';
-import generatePreFilledLink from '../utils/generatePreFilledLink';
+import getPreFilledLink from '../utils/getPreFilledLink';
 
 interface Props {
   resource: TResource;
@@ -10,7 +10,7 @@ interface Props {
 
 const UpdateResourceButton = ({ resource }: Props) => {
   return (
-    <Button href={generatePreFilledLink(resource)} target="_blank">
+    <Button href={getPreFilledLink(resource)} target="_blank">
       Request an Update
     </Button>
   );
